Convert to a sample-based decoder, allowing loop()
The libmad decoder library ddefault function runs until the entire MP3
is completed. This makes it unusable for event-driven things as you can't
do things like check for buttons, run a webserver, etc. while it is doing
the decode.

Build a AudioGeneratorMP3 class which calls internal libmad functions but
which supports returning after filling the I2S buffers, so it can be run
as part of the loop() without hogging the CPU.